+++ resolwe/flow/filters.py
@@ -687,5 +687,6 @@ class AnnotationValueFilter(BaseResolweFilter, metaclass=AnnotationValueFieldMet
                 "field__name": TEXT_LOOKUPS[:],
                 "field__label": TEXT_LOOKUPS[:],
                 "field__group__name": TEXT_LOOKUPS[:],
+                "modified": DATE_LOOKUPS[:],
             },
         }