+++ resolwe/flow/views/annotations.py
@@ -86,6 +86,7 @@ class AnnotationValueViewSet(
     filterset_class = AnnotationValueFilter
     queryset = AnnotationValue.objects.all()
     permission_classes = (get_permissions_class(),)
+    ordering_fields = ("modified",)
 
     def get_serializer(self, *args: Any, **kwargs: Any) -> BaseSerializer:
         """Get serializer instance depending on the request type."""